Two layer rendering with basic priority support
Timings in CUS42 are not ideal but they are working. layer B is processed in the first 4 pixels by CUS42 and layer B is processed in the second four pixels in CUS43. This is in order to tie the working timings in with the ^U prom chip that is driven by the layer select (CLK_2H) line