incorporated stub sprite subsystem
stub sprite subsystem with associated memory is now operational. We should be able to scrutinise how the CPU reads/writes from the work memory stored in the object ram. So far from initial simulations there is knly one point during a write operation where data, address and we lines coincide to stable values on the positive edge of 6M. Unfortunately there are two positive edges in the write window and so I do not know if this is at all representational of what would be happening in reality.

That said, things are definitely looking more predictable and working to the spec of the known chips involved.